refactor: update float chat components imports to use shadcn directory
