+++ features/chat/components/floating/floating-chat.tsx
@@ -3,13 +3,13 @@
 import { useState, useEffect, useRef, useCallback } from 'react';
 import { MessageCircle, X, Maximize2, GripVertical } from 'lucide-react';
 import Link from 'next/link';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/shadcn/button';
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from '@/components/ui/tooltip';
+} from '@/components/shadcn/tooltip';
 import { FloatingChatRoomList } from './floating-chat-room-list';
 import { FloatingChatRoomView } from './floating-chat-room-view';
